# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OWAAR CLOUD API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface UpdateResponse {
  success: boolean;
  output: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Setting {
  key: string;
  valueAr: string;
  valueEn: string;
}

export interface UpdateSettingRequest {
  valueAr: string;
  valueEn: string;
}

export type ServiceProvider =
  (typeof ServiceProvider)[keyof typeof ServiceProvider];

export const ServiceProvider = {
  coolify: "coolify",
  cloudron: "cloudron",
} as const;

export interface Service {
  id: number;
  nameAr: string;
  nameEn: string;
  descAr: string;
  descEn: string;
  icon: string;
  templateId: string;
  provider: ServiceProvider;
  createdAt: string;
}

export type CreateServiceRequestProvider =
  (typeof CreateServiceRequestProvider)[keyof typeof CreateServiceRequestProvider];

export const CreateServiceRequestProvider = {
  coolify: "coolify",
  cloudron: "cloudron",
} as const;

export interface CreateServiceRequest {
  nameAr: string;
  nameEn: string;
  descAr: string;
  descEn: string;
  icon: string;
  templateId: string;
  provider: CreateServiceRequestProvider;
}

export type UpdateServiceRequestProvider =
  (typeof UpdateServiceRequestProvider)[keyof typeof UpdateServiceRequestProvider];

export const UpdateServiceRequestProvider = {
  coolify: "coolify",
  cloudron: "cloudron",
} as const;

export interface UpdateServiceRequest {
  nameAr?: string;
  nameEn?: string;
  descAr?: string;
  descEn?: string;
  icon?: string;
  templateId?: string;
  provider?: UpdateServiceRequestProvider;
}

export type DeploymentStatus =
  (typeof DeploymentStatus)[keyof typeof DeploymentStatus];

export const DeploymentStatus = {
  deploying: "deploying",
  running: "running",
  stopped: "stopped",
  failed: "failed",
} as const;

export interface Deployment {
  id: number;
  userId: number;
  serviceId: number;
  status: DeploymentStatus;
  domainUrl: string;
  service?: Service;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeploymentRequest {
  serviceId: number;
}