# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OWAAR CLOUD API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: settings
    description: Site settings / branding
  - name: services
    description: Deployable services
  - name: deployments
    description: User deployments
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all site settings
      responses:
        "200":
          description: Settings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"

  /settings/{key}:
    put:
      operationId: updateSetting
      tags: [settings]
      summary: Update a setting (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingRequest"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all deployable services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments for current user (or all for admin)
      responses:
        "200":
          description: Deployments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Deploy a service (one-click deploy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDeployment
      tags: [deployments]
      summary: Delete/stop a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/update:
    post:
      operationId: triggerUpdate
      tags: [admin]
      summary: Trigger a git pull self-update (admin)
      responses:
        "200":
          description: Update triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    UpdateResponse:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
      required:
        - success
        - output

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt

    Setting:
      type: object
      properties:
        key:
          type: string
        valueAr:
          type: string
        valueEn:
          type: string
      required:
        - key
        - valueAr
        - valueEn

    UpdateSettingRequest:
      type: object
      properties:
        valueAr:
          type: string
        valueEn:
          type: string
      required:
        - valueAr
        - valueEn

    Service:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        icon:
          type: string
        templateId:
          type: string
        provider:
          type: string
          enum: [coolify, cloudron]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - descAr
        - descEn
        - icon
        - templateId
        - provider
        - createdAt

    CreateServiceRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        icon:
          type: string
        templateId:
          type: string
        provider:
          type: string
          enum: [coolify, cloudron]
      required:
        - nameAr
        - nameEn
        - descAr
        - descEn
        - icon
        - templateId
        - provider

    UpdateServiceRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        icon:
          type: string
        templateId:
          type: string
        provider:
          type: string
          enum: [coolify, cloudron]

    Deployment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        status:
          type: string
          enum: [deploying, running, stopped, failed]
        domainUrl:
          type: string
        service:
          $ref: "#/components/schemas/Service"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - status
        - domainUrl
        - createdAt
        - updatedAt

    CreateDeploymentRequest:
      type: object
      properties:
        serviceId:
          type: integer
      required:
        - serviceId