# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OWAAR CLOUD API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceProvider } from "./serviceProvider";

export interface Service {
  id: number;
  nameAr: string;
  nameEn: string;
  descAr: string;
  descEn: string;
  icon: string;
  templateId: string;
  provider: ServiceProvider;
  createdAt: Date;
}