# lib/api-zod/src/generated/types/updateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OWAAR CLOUD API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateResponse {
  success: boolean;
  output: string;
}